+++ changelog.md
@@ -16,6 +16,7 @@
 - Added new card statuses
 - Add reference and rejected reason to transaction and additional transaction statuses
 - Add support for digital wallets in cards
+- Add statements in FrolloSDK
 
 ### 3.9.0